requirements.txt
anthropic
anyio==4.4.0
appdirs==1.4.4
asgiref==3.8.1
blinker==1.8.2
certifi==2024.6.2
charset-normalizer==3.3.2
click==8.1.7
Flask==3.0.3
h11==0.14.0
htmltools==0.5.2
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
langdetect==1.0.9
linkify-it-py==2.0.3
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdit-py-plugins==0.4.1
mdurl==0.1.2
numpy==1.26.4
openai
packaging==24.0
pandas==2.2.2
prompt-toolkit==3.0.36
python-dotenv
python-dateutil==2.9.0.post0
python-multipart==0.0.9
pytz==2024.1
questionary==2.0.1
requests==2.32.3
sacremoses==0.1.1
shiny==0.10.2
sentencepiece==0.2.0
sentence-transformers==3.0.1
six==1.16.0
sniffio==1.3.1
starlette==0.37.2
transformers
typing_extensions==4.12.1
tzdata==2024.1
uc-micro-py==1.0.3
urllib3==2.2.1
uvicorn==0.30.1
watchfiles==0.22.0
wcwidth==0.2.13
websockets==12.0
Werkzeug==3.0.3
